- Enable the -parameter build option. - Put invalidClassTest into separate test class.
