+++ servlet/resteasy-servlet-spring-boot-starter-test/src/test/java/com/sample/app/test/LogbackTestApplicationListener.java
@@ -38,15 +38,15 @@ public class LogbackTestApplicationListener implements SmartApplicationListener 
             if (event == null || warningOrErrorFound) {
                 return;
             }
-//            Level level = event.getLevel();
-//            if ((level.equals(Level.WARN) || level.equals(Level.ERROR))
-//                    && !event.getMessage().startsWith(SCANNING_WARNING)
-//                    && !event.getMessage().startsWith("InetAddress.getLocalHost") // On MacOS Java 11 it sometimes generate relative warning so we ignore it.
-//                    && !event.getLoggerName().equals("org.apache.tomcat.util.modeler.Registry") // Tomcat generate relative warnings when the servers are started/stoped multiple times during tests.
-//                    && !event.getLevel().equals("org.springframework.core.LocalVariableTableParameterNameDiscoverer")) // thrown since https://github.com/spring-projects/spring-boot/releases/tag/v3.0.0 {
-//                warningOrErrorFound = true;
-//                Assert.fail(event.getFormattedMessage());
-//            }
+            Level level = event.getLevel();
+            if ((level.equals(Level.WARN) || level.equals(Level.ERROR))
+                    && !event.getMessage().startsWith(SCANNING_WARNING)
+                    && !event.getMessage().startsWith("InetAddress.getLocalHost") // On MacOS Java 11 it sometimes generate relative warning so we ignore it.
+                    && !event.getLoggerName().equals("org.apache.tomcat.util.modeler.Registry")) // Tomcat generate relative warnings when the servers are started/stoped multiple times during tests.
+            {
+                warningOrErrorFound = true;
+                Assert.fail(event.getFormattedMessage());
+            }
         }
     };
